requirements.txt
Flask
yfinance
pandas
numpy
plotly
requests
textblob
pytrends
tensorflow
scikit-learn
python-dotenv
streamlit 
transformers 
vandersentiment
torch
nltk


